feature: new view and api call for casts, refactor: types in meaning type files
