+++ src/composables/services/index.ts
@@ -1,1 +1,2 @@
 export { default as useMoviesService } from './movies'
+export { default as useConfigurationService } from './configuration'
