+++ src/types/movies.ts
@@ -1,5 +1,8 @@
 import type { Ref } from 'vue'
 
+import type { Casts } from './casts'
+import type { Company, Country, Language, Genre, apiGenreType } from './globals'
+
 export interface apiMovieType {
     page: number
     results: [Movie]
@@ -22,19 +25,6 @@ export interface Movie {
     video: boolean
     vote_average: number
     vote_count: number
-}
-
-export interface apiGenreType {
-    genres: [Genre]
-}
-
-export interface Genre {
-    id: number
-    name: string
-}
-
-export interface SelectItem extends Genre {
-    isActive: boolean
 }
 
 export interface apiMovieDetailsType {
@@ -82,58 +72,6 @@ export interface Video {
     site: string
     size: number
     type: string
-}
-
-export interface Company {
-    id: number
-    logo_path: string
-    name: string
-    origin_country: string
-}
-
-export interface Country {
-    iso_3166_1: string
-    name: string
-}
-
-export interface Language {
-    english_name: string
-    iso_639_1: string
-    name: string
-}
-
-export interface Casts {
-    cast: [Cast]
-    crew: [Crew]
-}
-
-export interface Cast {
-    adult: boolean
-    gender: number
-    id: number
-    known_for_department: string
-    name: string
-    original_name: string
-    popularity: number
-    profile_path: string
-    cast_id: number
-    character: string
-    credit_id: string
-    order: number
-}
-
-export interface Crew {
-    adult: boolean
-    gender: number
-    id: number
-    known_for_department: string
-    name: string
-    original_name: string
-    popularity: number
-    profile_path: string
-    credit_id: string
-    department: string
-    job: string
 }
 
 export interface MovieService {
